+++ src/Cool/Image/SaveImage.h
@@ -26,7 +26,7 @@ namespace Cool::ImageU {
     img::Size::DataType          width,
     img::Size::DataType          height,
     const void*                  data,
-    int                          channels_count,
+    size_t                       channels_count,
     bool                         flip_vertically = true
 ) -> bool;
